requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
python-dotenv==1.0.0
requests==2.31.0
python-docx==1.1.0
reportlab==3.6.13
beautifulsoup4==4.12.2
